Modify highlevel delete to not return any value
Starting from version 2.3.2, Milvus no longer includes the primary keys in the result
when processing the delete operation with expressions.
This change is due to the large amount of data involved.
The delete interface no longer returns any results.
If no exceptions are thrown, it indicates a successful deletion.
However, for backward compatibility, if the primary_keys returned from Milvus is not empty
the list of primary keys is still returned.

Signed-off-by: zhenshan.cao <zhenshan.cao@zilliz.com>