Add filter parameter to delete (#1779)
Commencing from Milvus 2.3.3, an enhanced functionality has been introduced to
facilitate data deletion based on expressive criteria.
In this context, the inclusion of the "filter" parameter serves as an entry point
for accessing this feature.

All rows that conform to the specified filter expression will be eliminated.

The deletion condition should either be a list of primary keys to be deleted
or a filter expression.
If both are specified or if none are specified, an exception will be thrown.

Signed-off-by: zhenshan.cao <zhenshan.cao@zilliz.com>