Fix timeout duration in nested functions
In `wait_for...` methods, the required state is continuosly polled in a
for loop. Since the parent `wait_for...` and the poll method use the
same value of timeout, the `wait_for...` call could technically run for
longer than the timeout (upto 2x as long if you get unlucky)

The timeout that is passed to the inner polling methods should be
`original_timeout - time_elapsed` instead. This way the parent
`wait_for...` will never run for longer than the timeout.

Signed-off-by: Ashwin Krishna Kumar <ashwin.kumar6@ibm.com>