feat: allow empty sparse row (#2291)
currently pymilvus supports the following sparse vector formats:

* scipy sparse classes
* list of tuples: `[(dim, val), (dim, val), ...]`
* dict: `{dim: val, dim: val, ...}`

When dealing with input data, pymilvus doesn't check the schema to infer
the data type, instead it tries to infer the data type based on the
shape. that's why we have `entity_is_sparse_matrix` method.

Now we want to support empty sparse vector row, it is easy for scipy
sparse classes, but for the other 2 format, it means we must infer empty
list/dict `[]/{}` as sparse vector.

Will this be problematic? what could be the other options?

Signed-off-by: Buqian Zheng <zhengbuqian@gmail.com>