fix: remove duplicate field name check in pymilvus (#2294)
this check was added in https://github.com/milvus-io/pymilvus/pull/2257.

this caused an test in milvus to fail. the test expects the error to be
thrown on server side instead of sdk side. it is ok to check on sdk side
but I choose to remove it to keep the behavior consistent and comfort
the ut without modifying it.

Failed test:
https://github.com/milvus-io/milvus/blob/5b2658327d676453e8a251ffc3273f4a218fb278/tests/python_client/testcases/test_collection.py#L318

Signed-off-by: Buqian Zheng <zhengbuqian@gmail.com>