Corrected grammar and consistency in error messages (#2289) (#2330)
This PR updates several error messages for grammar and clarity. The
changes include:

Corrected "data should be a list of list" to "The data should be a list
of lists" for better readability.
Fixed grammar in "The data don't match with schema fields" to "The data
doesn't match with schema fields."
Corrected field name mismatch message from "The name of field don't
match" to "The name of the field doesn't match."
These improvements enhance the clarity and professionalism of error
messages in the codebase.

Signed-off-by: zhenshan.cao <zhenshan.cao@zilliz.com>
Co-authored-by: Ahmet Yasin Aytar <34247619+Ahmetyasin@users.noreply.github.com>